Revert "Word cloud (#30)"
This reverts commit cae63310019b145a6cf1b717ad8ec6ea11cedfe4.
