--- app/modules/WordCloud/README.md
@@ -1,5 +1,0 @@
-# WordCloud
-
-一个用于生成词云的模块。
-
-开启后，会存储群聊消息，并在每天 23:59 生成本日词云图片。
