--- app/modules/WordCloud/data_manager.py
@@ -1,29 +1,0 @@
-import sqlite3
-import os
-from . import MODULE_NAME
-
-
-class DataManager:
-    def __init__(
-        self, group_id
-    ):  #  这里以群聊为例，如果需要处理私聊，可以传入user_id或其他实现方法
-        data_dir = os.path.join("data", MODULE_NAME)
-        os.makedirs(data_dir, exist_ok=True)
-        db_path = os.path.join(
-            data_dir, f"group_{group_id}.db"
-        )  # 这里以群聊为例，如果需要处理私聊，可以传入user_id或其他实现方法
-        self.conn = sqlite3.connect(db_path)
-        self.cursor = self.conn.cursor()
-        self._create_table()
-
-    def _create_table(self):
-        """建表函数，如果表不存在则创建"""
-        self.cursor.execute("""CREATE TABLE IF NOT EXISTS""")
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.conn.close()
-
-    # 其他函数，可直接使用，无需手动关闭数据库连接。
