+++ app/modules/template/data_manager.py
@@ -9,36 +9,21 @@ class DataManager:
     ):  #  这里以群聊为例，如果需要处理私聊，可以传入user_id或其他实现方法
         data_dir = os.path.join("data", MODULE_NAME)
         os.makedirs(data_dir, exist_ok=True)
-        self.db_path = os.path.join(
+        db_path = os.path.join(
             data_dir, f"group_{group_id}.db"
         )  # 这里以群聊为例，如果需要处理私聊，可以传入user_id或其他实现方法
+        self.conn = sqlite3.connect(db_path)
+        self.cursor = self.conn.cursor()
         self._create_table()
 
     def _create_table(self):
         """建表函数，如果表不存在则创建"""
-        with sqlite3.connect(self.db_path) as conn:
-            conn.execute(
-                """
-                CREATE TABLE IF NOT EXISTS data_table (
-                    id INTEGER PRIMARY KEY AUTOINCREMENT,
-                    data TEXT NOT NULL
-                )
-            """
-            )
+        self.cursor.execute("""CREATE TABLE IF NOT EXISTS""")
 
-    # 其他函数，使用sqlite3.connect(self.db_path) as conn:进行数据库操作
-    def insert_data(self, data):
-        with sqlite3.connect(self.db_path) as conn:
-            conn.execute("INSERT INTO data_table (data) VALUES (?)", (data,))
-            conn.commit()
+    def __enter__(self):
+        return self
 
-    def get_data(self):
-        with sqlite3.connect(self.db_path) as conn:
-            cursor = conn.execute("SELECT data FROM data_table")
-            data = cursor.fetchall()
-            return [row[0] for row in data]
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.conn.close()
 
-    def delete_data(self, id: int):
-        with sqlite3.connect(self.db_path) as conn:
-            conn.execute("DELETE FROM data_table WHERE id = ?", (id,))
-            conn.commit()
+    # 其他函数，可直接使用，无需手动关闭数据库连接。
