+++ requirements.txt
@@ -6,5 +6,3 @@ numpy
 scikit-learn
 jieba
 scipy
-matplotlib
-wordcloud
